# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get daily dashboard summary
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /inventory/items:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List warehouse inventory items
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: lowStockOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/items/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "404":
          description: Inventory item not found
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Inventory item not found
  /inventory/movements:
    get:
      operationId: listInventoryMovements
      tags: [inventory]
      summary: List inventory movements
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [in, out, kitchen, adjustment]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Inventory movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryMovement"
    post:
      operationId: createInventoryMovement
      tags: [inventory]
      summary: Record warehouse or kitchen movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryMovementInput"
      responses:
        "201":
          description: Created movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryMovement"
  /finance/expenses:
    get:
      operationId: listExpenses
      tags: [finance]
      summary: List expenses
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [finance]
      summary: Record an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /finance/expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [finance]
      summary: Update expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Expense not found
    delete:
      operationId: deleteExpense
      tags: [finance]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Expense not found
  /finance/income:
    get:
      operationId: listIncome
      tags: [finance]
      summary: List income entries
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Income entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Income"
    post:
      operationId: createIncome
      tags: [finance]
      summary: Record income
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeInput"
      responses:
        "201":
          description: Created income
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
  /finance/income/{id}:
    patch:
      operationId: updateIncome
      tags: [finance]
      summary: Update income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeUpdate"
      responses:
        "200":
          description: Updated income
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Income"
        "404":
          description: Income not found
    delete:
      operationId: deleteIncome
      tags: [finance]
      summary: Delete income
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Income not found
  /staff/employees:
    get:
      operationId: listEmployees
      tags: [staff]
      summary: List employees
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [staff]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /staff/employees/{id}:
    patch:
      operationId: updateEmployee
      tags: [staff]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Employee not found
    delete:
      operationId: deleteEmployee
      tags: [staff]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Employee not found
  /staff/attendance:
    get:
      operationId: listAttendance
      tags: [staff]
      summary: List attendance records
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: employeeId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [staff]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      required: [id, name, category, unit, currentStock, kitchenStock, minimumStock]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        unit: { type: string }
        currentStock: { type: number }
        kitchenStock: { type: number }
        minimumStock: { type: number }
        costPerUnit: { type: number }
        updatedAt: { type: string, format: date-time }
    InventoryItemInput:
      type: object
      required: [name, category, unit, currentStock, minimumStock]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        unit: { type: string, minLength: 1 }
        currentStock: { type: number, minimum: 0 }
        kitchenStock: { type: number, minimum: 0 }
        minimumStock: { type: number, minimum: 0 }
        costPerUnit: { type: number, minimum: 0 }
    InventoryItemUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        unit: { type: string, minLength: 1 }
        currentStock: { type: number, minimum: 0 }
        kitchenStock: { type: number, minimum: 0 }
        minimumStock: { type: number, minimum: 0 }
        costPerUnit: { type: number, minimum: 0 }
    InventoryMovement:
      type: object
      required: [id, itemId, itemName, type, quantity, unit, actor, createdAt]
      properties:
        id: { type: integer }
        itemId: { type: integer }
        itemName: { type: string }
        type:
          type: string
          enum: [in, out, kitchen, adjustment, transfer]
        quantity: { type: number }
        unit: { type: string }
        note: { type: ["string", "null"] }
        actor: { type: string }
        createdAt: { type: string, format: date-time }
...
    DashboardSummary:
      type: object
      required: [date, totalIncome, totalExpenses, netCash, lowStockCount, kitchenMovements, attendanceCount, recentActivity]
      properties:
        date: { type: string, format: date }
        totalIncome: { type: number }
        totalExpenses: { type: number }
        netCash: { type: number }
        lowStockCount: { type: integer }
        kitchenMovements: { type: integer }
        attendanceCount: { type: integer }
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
...